redundant casts: cleanup

+++ platform/platform-impl/src/com/intellij/ui/plaf/beg/IdeaMenuUI.java
@@ -73,7 +73,7 @@ public class IdeaMenuUI extends BasicMenuUI{
 
   protected void installDefaults() {
     super.installDefaults();
-    Integer integer = (Integer)UIUtil.getPropertyMaxGutterIconWidth(getPropertyPrefix());
+    Integer integer = UIUtil.getPropertyMaxGutterIconWidth(getPropertyPrefix());
     if (integer != null){
       myMaxGutterIconWidth = integer.intValue();
     }
